rename figure 4 to 3, figure 5 to 4
