--- RED/configs/example/Figure_5_RT3D_chemostat_prior.yaml
@@ -1,21 +1,0 @@
-defaults:
-  - /environment: chemostat
-  - /model: RT3D_agent
-  - _self_
-
-model:
-  val_learning_rate: 0.0001
-  pol_learning_rate: 0.00005
-  policy_act: sigmoid
-  max_length: 11
-  mem_size: 500_000_000
-  std: 0.1
-  noise_bounds: [-0.25, 0.25]
-  action_bounds: [0, 1]
-
-hidden_layer_size: [[64, 64], [128, 128]]
-test_episode: false
-policy_delay: 2
-max_std: 1
-explore_rate: "${.max_std}"
-save_path: results/
