+++ examples/Figure_4_RT3D_chemostat/train_RT3D.py
@@ -1,113 +1,116 @@
 
-import math
 import os
 import sys
-import time
 
 IMPORT_PATH = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.append(IMPORT_PATH)
-print(IMPORT_PATH)
 
-import multiprocessing
 
-import hydra
-import matplotlib
-import matplotlib.pyplot as plt
+import math
+
+import matplotlib as mpl
 import numpy as np
 from casadi import *
-from hydra.utils import instantiate
-from omegaconf import DictConfig, OmegaConf
 
-from RED.agents.continuous_agents.rt3d import RT3D_agent
+mpl.use('tkagg')
+import json
+import multiprocessing
+import time
+
+import hydra
+import matplotlib.pyplot as plt
+import tensorflow as tf
+from hydra.utils import instantiate
+from omegaconf import DictConfig
+
+from RED.agents.continuous_agents import RT3D_agent
 from RED.environments.chemostat.xdot_chemostat import xdot
 from RED.environments.OED_env import OED_env
 
-OmegaConf.register_new_resolver("eval", eval) # https://omegaconf.readthedocs.io/en/2.3_branch/how_to_guides.html#how-to-perform-arithmetic-using-eval-as-a-resolver
-
 
 @hydra.main(version_base=None, config_path="../../RED/configs", config_name="example/Figure_4_RT3D_chemostat")
-def train_RT3D(cfg : DictConfig):
-    ### config setup
+def train_RT3D_prior(cfg : DictConfig):
+    #setup
     cfg = cfg.example
-    print(
-        "--- Configuration ---",
-        OmegaConf.to_yaml(cfg, resolve=True),
-        "--- End of configuration ---",
-        sep="\n\n"
+    
+    actual_params = DM(cfg.environment.actual_params)
+    normaliser = np.array(cfg.environment.normaliser)
+    n_params = actual_params.size()[0]
+    n_system_variables = len(cfg.environment.y0)
+    n_FIM_elements = sum(range(n_params + 1))
+    n_tot = n_system_variables + n_params * n_system_variables + n_FIM_elements
+    param_guesses = actual_params
+    physical_devices = tf.config.list_physical_devices('GPU')
+    
+    n_cores = multiprocessing.cpu_count()
+    try:
+        tf.config.experimental.set_memory_growth(physical_devices[0], True)
+    except:
+        pass
+
+    #agent setup
+    pol_layer_sizes = [cfg.environment.n_observed_variables + 1, cfg.environment.n_observed_variables + 1 + cfg.environment.n_controlled_inputs, cfg.hidden_layer_size[0], cfg.hidden_layer_size[1], cfg.environment.n_controlled_inputs]
+    val_layer_sizes = [cfg.environment.n_observed_variables + 1 + cfg.environment.n_controlled_inputs, cfg.environment.n_observed_variables + 1 + cfg.environment.n_controlled_inputs, cfg.hidden_layer_size[0], cfg.hidden_layer_size[1], 1]
+
+    agent = instantiate(
+        cfg.model,
+        val_layer_sizes=val_layer_sizes,
+        pol_layer_sizes=pol_layer_sizes,
+        batch_size=int(cfg.environment.N_control_intervals * cfg.environment.skip),
+        max_length=11,
     )
 
-    ### prepare save path
-    save_path = os.path.join(cfg.save_path, time.strftime("%Y-%m-%d_%H-%M"))
-    os.makedirs(save_path, exist_ok=True)
-    print("Results will be saved in: ", save_path)
-
-    ### agent setup
-    agent = instantiate(cfg.model)
-    explore_rate = cfg.initial_explore_rate
-    seq_dim = cfg.environment.n_observed_variables + 1 + cfg.environment.n_controlled_inputs
-
-    ### env setup
-    env, n_params = setup_env(cfg)
-    
-    total_episodes = cfg.environment.n_episodes // cfg.environment.skip
-    history = {k: [] for k in ["returns", "actions", "rewards", "us", "explore_rate"]}
+    explore_rate = cfg.explore_rate
     update_count = 0
+    alpha = 1
+    all_returns = []
+    all_test_returns = []
 
-    ### training loop
-    for episode in range(total_episodes):
-        actual_params = np.random.uniform(
-            low=cfg.environment.actual_params,
-            high=cfg.environment.actual_params,
-            size=(cfg.environment.skip, n_params)
-        )
+    #env setup
+    args = cfg.environment.y0, xdot, param_guesses, actual_params, cfg.environment.n_observed_variables, \
+        cfg.environment.n_controlled_inputs, cfg.environment.num_inputs, cfg.environment.input_bounds, \
+        cfg.environment.dt, cfg.environment.control_interval_time, normaliser
+    env = OED_env(*args)
+    env.mapped_trajectory_solver = env.CI_solver.map(cfg.environment.skip, "thread", n_cores)
+
+    for episode in range(int(cfg.environment.n_episodes//cfg.environment.skip)): #training loop
+
+        actual_params = np.random.uniform(low=cfg.environment.lb, high=cfg.environment.ub,  size = (cfg.environment.skip, 3)) # sample from uniform distribution
         env.param_guesses = DM(actual_params)
-        
-        ### episode buffers for agent
         states = [env.get_initial_RL_state_parallel() for i in range(cfg.environment.skip)]
-        trajectories = [[] for _ in range(cfg.environment.skip)]
-        sequences = [[[0] * seq_dim] for _ in range(cfg.environment.skip)]
-
-        ### episode logging buffers
         e_returns = [0 for _ in range(cfg.environment.skip)]
         e_actions = []
+        e_exploit_flags =[]
         e_rewards = [[] for _ in range(cfg.environment.skip)]
         e_us = [[] for _ in range(cfg.environment.skip)]
-
-        ### reset env between episodes
+        trajectories = [[] for _ in range(cfg.environment.skip)]
+        sequences = [[[0]*pol_layer_sizes[1]] for _ in range(cfg.environment.skip)]
         env.reset()
         env.param_guesses = DM(actual_params)
         env.logdetFIMs = [[] for _ in range(cfg.environment.skip)]
         env.detFIMs = [[] for _ in range(cfg.environment.skip)]
 
-        ### run an episode
-        for control_interval in range(0, cfg.environment.N_control_intervals):
+        for e in range(0, cfg.environment.N_control_intervals): # run an episode
             inputs = [states, sequences]
-
-            ### get agent's actions
             if episode < 1000 // cfg.environment.skip:
-                actions = agent.get_actions(inputs, explore_rate=1, test_episode=True, recurrent=True)
+                actions = agent.get_actions(inputs, explore_rate = 1, test_episode=cfg.test_episode, recurrent=True)
             else:
-                actions = agent.get_actions(inputs, explore_rate=explore_rate, test_episode=True, recurrent=True)
+                actions = agent.get_actions(inputs, explore_rate=explore_rate, test_episode=cfg.test_episode, recurrent=True)
 
             e_actions.append(actions)
-            outputs = env.map_parallel_step(actions.T, actual_params, continuous = True)
+            outputs = env.map_parallel_step(np.array(actions).T, actual_params, continuous = True)
             next_states = []
 
-            for i, obs in enumerate(outputs): #extract outputs from experiments run in parallel
-                state = states[i]
-                action = actions[i]
-                
-                next_state, reward, done, _, u  = obs
+            for i,o in enumerate(outputs): #extract outputs from parallel experiments
+                next_state, reward, done, _, u  = o
                 e_us[i].append(u)
                 next_states.append(next_state)
+                state = states[i]
+                action = actions[i]
 
-                ### set done flag
-                if control_interval == cfg.environment.N_control_intervals - 1 \
-                    or np.all(np.abs(next_state) >= 1) \
-                    or math.isnan(np.sum(next_state)):
+                if e == cfg.environment.N_control_intervals - 1 or np.all(np.abs(next_state) >= 1) or math.isnan(np.sum(next_state)):
                     done = True
 
-                ### memorize transition
                 transition = (state, action, reward, next_state, done)
                 trajectories[i].append(transition)
                 sequences[i].append(np.concatenate((state, action)))
@@ -116,63 +119,63 @@ def train_RT3D(cfg : DictConfig):
                     e_returns[i] += reward
             states = next_states
 
-        ### append trajectories to memory
+        if cfg.test_episode:
+            trajectories = trajectories[:-1]
         for trajectory in trajectories:
-            # check for instability
-            if np.all([np.all(np.abs(trajectory[i][0]) <= 1) for i in range(len(trajectory))]) \
-                and not math.isnan(np.sum(trajectory[-1][0])):
+            if np.all( [np.all(np.abs(trajectory[i][0]) <= 1) for i in range(len(trajectory))] ) and not math.isnan(np.sum(trajectory[-1][0])): # check for instability
                 agent.memory.append(trajectory)
 
-        ### train agent
-        if episode > 1000 // cfg.environment.skip:
+
+        if episode > 1000 // cfg.environment.skip: #train agent
+            print('training', update_count)
+            t = time.time()
             for _ in range(cfg.environment.skip):
+
                 update_count += 1
                 policy = update_count % cfg.policy_delay == 0
-                agent.Q_update(policy=policy, recurrent=True)
 
-        ### update explore rate
-        explore_rate = cfg.max_std * agent.get_rate(
-            episode=episode,
-            min_rate=0,
-            max_rate=1,
-            denominator=cfg.environment.n_episodes / (11 * cfg.environment.skip)
-        )
+                agent.Q_update(policy=policy, fitted=False, recurrent=True, low_mem=False)
+            print('fitting time', time.time() - t)
 
-        ### log results
-        history["returns"].extend(e_returns)
-        history["actions"].extend(e_actions)
-        history["rewards"].extend(e_rewards)
-        history["us"].extend(e_us)
-        history["explore_rate"].append(explore_rate)
+        explore_rate = agent.get_rate(episode, 0, 1, cfg.environment.n_episodes / (11 * cfg.environment.skip)) * cfg.max_std
 
+        if cfg.test_episode:
+            all_returns.extend(e_returns[:-1])
+            all_test_returns.append(np.sum(np.array(e_rewards)[-1, :]))
+        else:
+            all_returns.extend(e_returns)
+        all_returns.extend(e_returns)
         print()
-        print(f"EPISODE: [{episode}/{total_episodes}] ({episode * cfg.environment.skip})")
-        print(f"explore rate: {explore_rate}")
-        print(f"average return: {np.mean(history['returns'][-cfg.environment.skip:])}")
+        print('EPISODE: ', episode, episode*cfg.environment.skip)
 
-    ### save results and plot
-    agent.save_network(save_path)
-    for k in history.keys():
-        np.save(os.path.join(save_path, f'{k}.npy'), np.array(history[k]))
+
+        print('av return: ', np.mean(all_returns[-cfg.environment.skip:]))
+        print()
+
+        if cfg.test_episode:
+            print('test actions:', np.array(e_actions)[:, -1])
+            print('test rewards:', np.array(e_rewards)[-1, :])
+            print('test return:', np.sum(np.array(e_rewards)[-1, :]))
+            print()
+
+    #plot and save results
+    agent.save_network(cfg.save_path)
+    np.save(os.path.join(cfg.save_path, 'all_returns.npy'), np.array(all_returns))
+    np.save(os.path.join(cfg.save_path, 'actions.npy'), np.array(agent.actions))
+
+    if cfg.test_episode:
+        np.save(cfg.save_path + '/all_test_returns.npy', np.array(all_test_returns))
+
 
     t = np.arange(cfg.environment.N_control_intervals) * int(cfg.environment.control_interval_time)
-    plt.plot(history['returns'])
-    plt.show()
+
+    print(all_returns)
+    print(agent.actions)
 
 
-def setup_env(cfg):
-    n_cores = multiprocessing.cpu_count()
-    actual_params = DM(cfg.environment.actual_params)
-    normaliser = np.array(cfg.environment.normaliser)
-    n_params = actual_params.size()[0]
-    param_guesses = actual_params
-    args = cfg.environment.y0, xdot, param_guesses, actual_params, cfg.environment.n_observed_variables, \
-        cfg.environment.n_controlled_inputs, cfg.environment.num_inputs, cfg.environment.input_bounds, \
-        cfg.environment.dt, cfg.environment.control_interval_time, normaliser
-    env = OED_env(*args)
-    env.mapped_trajectory_solver = env.CI_solver.map(cfg.environment.skip, "thread", n_cores)
-    return env, n_params
+    #plt.plot(all_returns)
+    #plt.show()
 
 
 if __name__ == '__main__':
-    train_RT3D()
+    train_RT3D_prior()
